requirements.txt
crewai==0.28.8
mcp==1.0.0
flask==3.0.0
pymongo==4.6.1
qdrant-client==1.7.0
mistralai==0.1.2
openai==1.12.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
pillow==10.2.0
tweepy==4.14.0
facebook-sdk==3.1.0
linkedin-api==2.0.0
instagram-basic-display-api==2.0.0